+++ app/src/main/java/com/johnkasun/patch_matcher/EvaluationActivity.java
@@ -15,7 +15,7 @@ public class EvaluationActivity extends AppCompatActivity {
         setContentView(R.layout.activity_evaluation);
 
         Intent intent = getIntent();
-        float iEvaluationValue = intent.getIntExtra("Result", -1);
+        float iEvaluationValue = intent.getFloatExtra("Result", -1);
         TextView scoreView = findViewById(R.id.ScoreView);
         scoreView.setText(iEvaluationValue+"%");
     }
